break cyclic dependency between Money and Money.Allocation via extension methods
