+++ src/NMoneys/NMoneys.csproj
@@ -46,10 +46,12 @@
     <Compile Include="Allocations\EvenAllocator.cs" />
     <Compile Include="Allocations\FirstToLastAllocator.cs" />
     <Compile Include="Allocations\IRemainderAllocator.cs" />
+    <Compile Include="Allocations\Operations.Allocate.cs" />
     <Compile Include="Allocations\ProRataAllocator.cs" />
     <Compile Include="Allocations\RandomAllocator.cs" />
     <Compile Include="Allocations\Ratio.cs" />
     <Compile Include="Allocations\RatioCollection.cs" />
+    <Compile Include="Allocations\RemainderAllocator.cs" />
     <Compile Include="Allocations\RemainderAllocatorBase.cs" />
     <Compile Include="Allocations\LastToFirstAllocator.cs" />
     <Compile Include="Change\ChangeSolution.cs" />
@@ -74,7 +76,6 @@
     <Compile Include="Money.Serialization.Xml.net.cs" />
     <Compile Include="Money.Serialization.net.cs" />
     <Compile Include="Money.Arithmetic.cs" />
-    <Compile Include="Money.Allocations.cs" />
     <Compile Include="Money.Arithmetic.Operators.cs" />
     <Compile Include="Money.Parsing.cs" />
     <Compile Include="Money.Cloning.net.cs" />
@@ -97,7 +98,6 @@
     <Compile Include="ObsoleteCurrencyEventBehavior.cs" />
     <Compile Include="Properties\NMoneys_AssemblyInfo.cs" />
     <Compile Include="Providers.cs" />
-    <Compile Include="RemainderAllocator.cs" />
     <Compile Include="Serialization\Data.net.cs" />
     <Compile Include="Serialization\DataContractSurrogate.net.cs" />
     <Compile Include="Support\Comparer.polyfill.cs" />
